requirements.txt
arviz==0.15.1
DateTime==4.4
matplotlib==3.7.1
numpy==1.24.4
pandas==1.5.3
pyportfolioopt==1.5.5
pytensor==2.17.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pymc==5.9.0
QuantStats==0.0.62
requests==2.28.1
seaborn==0.11.1
scikit-learn==1.3.1
scipy==1.10.1
xarray==2023.1.0
yfinance==0.1.74
IPython==8.18.1